Move VT setup to sddm-helper
Previously, sddm set up the virtual terminal and jumped to it and after that,
sddm-helper opens it again and takes control of it. This only worked because
sddm leaked the terminal fd into sddm-helper.

The right way is to set up the virtual terminal and take control of it before
jumping to it. This is achieved by moving this responsibility to sddm-helper.

As the race is now fixed, also avoid leaking FDs in jumpToVt now.

Fixes #1078
Fixes #1083