Allow addition env vars to be defined in session files (#1370)
We have a continual problem with developers using Plasma in custom
prefixes. We need to set mulitple additional envs and it is a constant
pain point.

Setting XDG_DATA_DIRS and alike in a script in the Exec line of the
desktop file doesn't work properly, as this is after systemd and dbus
have already been started during pam_logind. It is currently very bodged
with more manual steps.

Things will work much better if we can set the env before pam_logind is
run.

Using pam_env or profile.d doesn't work as that affects all the
sessions.

It is not intended to be used for anything other than dev setups.